backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
google-generativeai==0.8.3
python-dotenv==1.0.1
httpx==0.27.2
ics==0.7.2
qrcode[pil]==7.4.2
